requirements.txt
Django>=5.2.7
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
requests>=2.31.0
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.1.0
